Split remaining serializers into their own files.
