+++ mathesar/api/serializers/schemas.py
@@ -1,0 +1,22 @@
+from rest_framework import serializers
+
+from mathesar.models import Schema
+
+
+class ModelNameField(serializers.CharField):
+    """
+    De-serializes the request field as a string, but serializes the response field as
+    `model.name`. Required to support passing and returing a model name from the
+    endpoint, while also storing the model as a related field.
+    """
+    def to_representation(self, value):
+        return value.name
+
+
+class SchemaSerializer(serializers.HyperlinkedModelSerializer):
+    name = serializers.CharField()
+    database = ModelNameField(max_length=128)
+
+    class Meta:
+        model = Schema
+        fields = ['id', 'name', 'database', 'has_dependencies']
